toolkit: gather all the repos required by the toolkit in a repo group
This enables people (and the CI) to fetch only the minimal subset of
projects that are required to make the CLIP OS toolkit work and thus
to avoid fetching unneeded repositories (which lead to a significant
quicker sync). Here is the procedure:

  $ repo init -g toolkit  # you may specify -u (for URL) or -b
  $ repo sync

For more details, see the documenation about the "groups" attribute in
the manifest format specification with `repo help manifest`.

Change-Id: I3fd33830a14cde87ca53c040487cbf5314db0320