manifest: delete README-FIRST and LICENSE files at root of source tree
The presence of a LGPL 2.1+ LICENSE file at the root of the CLIP OS
source tree may imply that the complete source tree is licensed under
LGPL 2.1+ which is not the case. Deleting this.

README-FIRST indicates to the user that the CLIP OS source tree has
been succesfully fetched and synchronized which may be false in the
case where the user has forgotten to setup the Git LFS filters.
Her/his source tree would contain in that case Git repositories with
missing Git LFS objects. Let's delete this README-FIRST file.

Change-Id: I6147dd1b7716ea96e43ec1c780c36c9b4b09321e